feat(modeling): unify expand into offset
