--- packages/modeling/src/operations/expansions/expand.js
@@ -1,40 +1,0 @@
-import { flatten } from '../../utils/flatten.js'
-
-import * as geom2 from '../../geometries/geom2/index.js'
-import * as geom3 from '../../geometries/geom3/index.js'
-import * as path2 from '../../geometries/path2/index.js'
-
-import { expandGeom2 } from './expandGeom2.js'
-import { expandGeom3 } from './expandGeom3.js'
-import { expandPath2 } from './expandPath2.js'
-
-/**
- * Expand the given geometry using the given options.
- * Both internal and external space is expanded for 2D and 3D shapes.
- *
- * Note: Contract is expand using a negative delta.
- * @param {Object} options - options for expand
- * @param {Number} [options.delta=1] - delta (+/-) of expansion
- * @param {String} [options.corners='edge'] - type of corner to create after expanding; edge, chamfer, round
- * @param {Integer} [options.segments=16] - number of segments when creating round corners
- * @param {...Objects} objects - the geometries to expand
- * @return {Object|Array} new geometry, or list of new geometries
- * @alias module:modeling/expansions.expand
- *
- * @example
- * let newArc = expand({delta: 5, corners: 'edge'}, arc({}))
- * let newSquare = expand({delta: 5, corners: 'chamfer'}, square({size: 30}))
- * let newSphere = expand({delta: 2, corners: 'round'}, cuboid({size: [20, 25, 5]}))
- */
-export const expand = (options, ...objects) => {
-  objects = flatten(objects)
-  if (objects.length === 0) throw new Error('wrong number of arguments')
-
-  const results = objects.map((object) => {
-    if (path2.isA(object)) return expandPath2(options, object)
-    if (geom2.isA(object)) return expandGeom2(options, object)
-    if (geom3.isA(object)) return expandGeom3(options, object)
-    return object
-  })
-  return results.length === 1 ? results[0] : results
-}
