+++ packages/modeling/src/operations/expansions/index.d.ts
@@ -1,4 +1,4 @@
-export { expand, ExpandOptions } from './expand'
 export { offset, OffsetOptions } from './offset'
+export { offsetFromPoints } from './offset'
 
 export as namespace expansions
