+++ packages/modeling/src/operations/expansions/offsetFromPoints.d.ts
@@ -1,0 +1,10 @@
+import { Vec2 } from '../../maths/vec2/type'
+
+export interface OffsetFromPointsOptions {
+  delta?: number
+  corners?: 'edge' | 'chamfer' | 'round'
+  segments?: number
+  closed?: boolean
+}
+
+export function offsetFromPoints(options?: OffsetFromPointsOptions, points: Array<Vec2>): Array<Vec2>
