+++ packages/modeling/src/operations/expansions/offsetGeom2.test.js
@@ -4,30 +4,55 @@ import { geom2 } from '../../geometries/index.js'
 
 import { measureArea } from '../../measurements/index.js'
 
-import { roundedRectangle } from '../../primitives/index.js'
+import { roundedRectangle, square } from '../../primitives/index.js'
 
 import { offset } from './index.js'
 
 import { comparePoints } from '../../../test/helpers/index.js'
+
+test('offset: offset an empty geom2', (t) => {
+  const empty = geom2.create()
+  const obs = offset({ delta: 1 }, empty)
+  const pts = geom2.toPoints(obs)
+  const exp = []
+  t.notThrows(() => geom2.validate(obs))
+  t.is(measureArea(obs), 0)
+  t.true(comparePoints(pts, exp))
+})
+
+test('offset: offset of a geom2 produces expected changes to points', (t) => {
+  const geometry = square({ size: 16 })
+
+  const obs = offset({ delta: 2, corners: 'round', segments: 8 }, geometry)
+  const pts = geom2.toPoints(obs)
+  const exp = [
+    [-9.414213562373096, -9.414213562373096],
+    [-8, -10],
+    [8, -10],
+    [9.414213562373096, -9.414213562373096],
+    [10, -8],
+    [10, 8],
+    [9.414213562373096, 9.414213562373096],
+    [8, 10],
+    [-8, 10],
+    [-9.414213562373096, 9.414213562373096],
+    [-10, 8],
+    [-10, -8]
+  ]
+  t.notThrows(() => geom2.validate(obs))
+  t.is(pts.length, 12)
+  t.true(comparePoints(pts, exp))
+})
 
 test('offset (options): offsetting of a simple geom2 produces expected offset geom2', (t) => {
   const geometry = geom2.create([
     [[-5, -5], [5, -5], [5, 5], [3, 5], [3, 0], [-3, 0], [-3, 5], [-5, 5]]
   ])
 
-  // empty
-  const empty = geom2.create()
-  let obs = offset({ delta: 1 }, empty)
-  let pts = geom2.toPoints(obs)
-  let exp = []
-  t.notThrows(() => geom2.validate(obs))
-  t.is(measureArea(obs), 0)
-  t.true(comparePoints(pts, exp))
-
   // expand +
-  obs = offset({ delta: 1, corners: 'round' }, geometry)
-  pts = geom2.toPoints(obs)
-  exp = [
+  let obs = offset({ delta: 1, corners: 'round', segments: 0 }, geometry)
+  let pts = geom2.toPoints(obs)
+  let exp = [
     [-5, -6],
     [5, -6],
     [6, -5],
@@ -49,7 +74,7 @@ test('offset (options): offsetting of a simple geom2 produces expected offset ge
   t.true(comparePoints(pts, exp))
 
   // contract -
-  obs = offset({ delta: -0.5, corners: 'round' }, geometry)
+  obs = offset({ delta: -0.5, corners: 'round', segments: 0 }, geometry)
   pts = geom2.toPoints(obs)
   exp = [
     [-4.5, -4.5],
@@ -139,7 +164,7 @@ test('offset (options): offsetting of a complex geom2 produces expected offset g
   ])
 
   // expand +
-  const obs = offset({ delta: 2, corners: 'edge' }, geometry)
+  const obs = offset({ delta: 2, corners: 'edge', expandHoles: true }, geometry)
   const pts = geom2.toPoints(obs)
   const exp = [
     [77, -77],
